feat: enable usage of an extra disk
vmware customization script cannot prevent /usr/sbin/cirb-extend-storage
to be executed. cirb-extend-storage extend the lv /srv based on the
space left in the main VG. IOW if we add more disk space into the VG, the space
left will be automaticaly allocated to /srv.

Sometimes we want to control the disk layout and create other lv within
this VG.
For the moment, the only way is to add an extra disk and register using
the customization script.

This change enable the user to declare the size of this extra disk.

refs CICDITS-692